consistency cleanup for headers, file names
