+++ Sources/Chart/Chart.swift
@@ -1,9 +1,6 @@
 //
 //  Chart.swift
 //
-//
-//  Created by Joseph Heck on 3/25/22.
-//
 
 import SwiftUI
