+++ Sources/Chart/ChartBuilder.swift
@@ -1,9 +1,6 @@
 //
 //  MarkBuilder.swift
 //
-//
-//  Created by Joseph Heck on 4/7/22.
-//
 
 import Foundation
