+++ Sources/Chart/Marks/VisualChannel.swift
@@ -1,9 +1,6 @@
 //
 //  VisualChannel.swift
 //
-//
-//  Created by Joseph Heck on 3/25/22.
-//
 
 import CoreGraphics
 import Foundation
